requirements.txt
# direct dependencies
bcrypt==4.0.1
cachetools==5.2.1
fido2==1.1.0
Flask==2.2.2
Flask-Login==0.6.2
tinydb==4.7.1
gunicorn==20.1.0
flask-migrate
sqlalchemy
pymysql
jsonify


# indirect dependencies
#cffi==1.15.1
click==8.1.3
cryptography==39.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
Werkzeug==2.2.2
